% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/slidify.R
\name{check_slidifyLibraries}
\alias{check_slidifyLibraries}
\title{Convert an Rmd document into HTML5 using a framework}
\usage{
check_slidifyLibraries()
}
\arguments{
\item{inputFile}{path to input Rmd document}

\item{outputFile}{path to output html document; default uses inputFile}

\item{knit_deck}{should the input file be knit?; default is TRUE}
}
\value{
path to outputFile
}
\description{
Convert an Rmd document into HTML5 using a framework
}
\seealso{
slidify-package
}

