\name{publish_dropbox}
\alias{publish_dropbox}
\title{Publish slide deck to Dropbox}
\usage{
  publish_dropbox(dirname)
}
\arguments{
  \item{dirname}{name of directory to publish to; defaults
  to slide directory}
}
\description{
  Publish slide deck to Dropbox
}
\seealso{
  Other publish: \code{\link{publish}},
  \code{\link{publish_github}}
}

