% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/publish.R
\name{publish_gist}
\alias{publish_gist}
\title{Publish slide deck to gist}
\usage{
publish_gist(title, filenames = dir(".", pattern = "index"), public = T)
}
\arguments{
\item{title}{title of the presentation}

\item{files}{list of files to publish, defaults to index.* files}
}
\description{
Publish slide deck to gist
}

