% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/render.R
\name{render_page}
\alias{render_page}
\title{Render page}
\usage{
render_page(page, payload, return_page = FALSE, save_payload = FALSE)
}
\arguments{
\item{page}{list containing the parsed page}

\item{payload}{list containing site and pages}

\item{return_page}{return the page object after rendering}

\item{save_payload}{save the payload as a .Rdata file?}
}
\description{
Render page
}

