% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{runDeck}
\alias{runDeck}
\title{Run a slide Deck}
\usage{
runDeck(deckDir = ".", appDir = file.path(deckDir, "apps"), shiny = TRUE,
  ...)
}
\arguments{
\item{deckDir}{directory where the deck resides}

\item{appDir}{directory for shiny app}

\item{shiny}{use shiny}
}
\description{
Run a slide deck
}

