\docType{package}
\name{slidify-package}
\alias{slidify}
\alias{slidify-package}
\title{Create stunning presentations from R Markdown}
\description{
  Slidify is an R package to create, customize and publish
  HTML5 slide decks from R Markdown. See the package
  website in the references for more information and
  examples.
}
\author{
  Ramnath Vaidyanathan <\url{http://www.slidify.org}>
}
\references{
  Full documentation and demos:
  \url{http://www.slidify.org}; FAQ's:
  \url{https://github.com/ramnathv/slidify/blob/master/FAQ.md}
}
\seealso{
  The core function in this package: \code{\link{slidify}}
}

