\name{slidify}
\alias{slidify}
\title{Convert an Rmd document into HTML5 using a framework}
\usage{
  slidify(inputFile, outputFile, knit_deck = TRUE)
}
\arguments{
  \item{inputFile}{path to input Rmd document}

  \item{outputFile}{path to output html document; default
  uses inputFile}

  \item{knit_deck}{should the input file be knit?; default
  is TRUE}
}
\value{
  path to outputFile
}
\description{
  Convert an Rmd document into HTML5 using a framework
}
\seealso{
  slidify-package
}

