% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/publish.R
\name{publish_rpubs}
\alias{publish_rpubs}
\title{Publish slide deck to rPubs}
\usage{
publish_rpubs(title, html_file = "index.html")
}
\arguments{
\item{title}{title of the presentation}

\item{html_file}{path to html file to publish; defaults to index.html}
}
\description{
Publish slide deck to rPubs
}

