% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/slidify.R
\name{slidify}
\alias{slidify}
\title{Convert an Rmd document into HTML5}
\usage{
slidify(inputFile, knit_deck = TRUE, return_page = FALSE,
  save_payload = FALSE, envir = parent.frame())
}
\arguments{
\item{inputFile}{path to the Rmd file to slidify}

\item{knit_deck}{whether the file needs to be run through knit}

\item{return_page}{should the function return the payload}

\item{save_payload}{should the payload be saved to the slide directory}

\item{envir}{is the environment in which to execute code.}
}
\description{
Convert an Rmd document into HTML5
}

